LPD-93373 Move package from background.task to backgroundtask (as the interface that is implementing is located in *.backgroundtask package).
There are also other classes located in *.background.task packages, but I think that is a mistake
