+++ modules/apps/portal-instances/portal-instances-api/src/main/java/com/liferay/portal/instances/backgroundtask/constants/PortalInstanceBackgroundTaskConstants.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.portal.instances.background.task.constants;
+package com.liferay.portal.instances.backgroundtask.constants;
 
 /**
  * @author Luis Ortiz
