+++ modules/apps/portal-instances/portal-instances-test/src/testIntegration/java/com/liferay/portal/instances/internal/backgroundtask/test/AddPortalInstanceBackgroundTaskExecutorTest.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.portal.instances.internal.background.task.test;
+package com.liferay.portal.instances.internal.backgroundtask.test;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
 import com.liferay.petra.string.StringPool;
@@ -192,7 +192,7 @@ public class AddPortalInstanceBackgroundTaskExecutorTest {
 				TestPropsValues.getUserId(),
 				BackgroundTaskConstants.GROUP_ID_DEFAULT,
 				"AddPortalInstance#" + _WEB_ID,
-				"com.liferay.portal.instances.internal.background.task." +
+				"com.liferay.portal.instances.internal.backgroundtask." +
 					"AddPortalInstanceBackgroundTaskExecutor",
 				taskContextMap, new ServiceContext());
